FunCoup/requirements.txt
Django==4.1.1
gunicorn==21.2.0
six==1.16.0
pyyaml==6.0
networkx==3.0
statsmodels==0.14.0
scipy==1.10.0
joblib==1.2.0
numpy==1.24.1
pandas==1.5.2
tqdm==4.65.0
matplotlib==3.7.1
django-postgres-copy==2.7.2
psycopg2-binary==2.9.5
scikit-learn==1.2.2